Registration controller
