+++ src/main/java/com/pawntracker/controller/UserController.java
@@ -38,82 +38,6 @@ public class UserController {
     }
 
 
-    @GetMapping("/registration")
-    public String registration(Model model) {
-        model.addAttribute("user", new User());
-
-        return "registration";
-    }
-
-    @PostMapping("/registration")
-    public String registration(@ModelAttribute("user") @Valid  User user, BindingResult bindingResult ,
-                               @RequestParam("file") MultipartFile file) throws IOException {
-
-        userValidator.validate(user,bindingResult);
-        final String rawPassword = user.getPassword();
-        if (bindingResult.hasErrors()) {
-            System.out.println("errors" + bindingResult.getAllErrors());
-            return "registration";
-        }
-        User user1 = userService.addImage(user, file);
-        userService.saveUserOrUpdate(user1);
-        securityService.autoLogin(user.getUsername(), rawPassword);
-
-        return "redirect:/registration/second_stage";
-    }
-
-    @GetMapping("/registration/second_stage")
-    public String secondStageRegistration(Model model) {
-        model.addAttribute("address", new Address());
-        model.addAttribute("phoneNumber", new PhoneNumber());
-        return "registration/second_stage";
-    }
-
-    @PostMapping("/registration/second_stage")
-    public String secondStageRegistration(Principal principal, @Valid Address address, @Valid PhoneNumber phoneNumber,
-                                          BindingResult result) {
-        if (result.hasErrors()) {
-            System.out.println("errors" + result.getAllErrors());
-            return "registration/second_stage";
-        }
-        userService.addAddresAndPhoneNumberToUser(address, phoneNumber, principal.getName());
-        return "redirect:/registration/choose-type-of-document";
-    }
-
-    @GetMapping("/registration/choose-type-of-document")
-     public String chooseDocument() {
-        return "registration/choose_document";
-    }
-
-    @GetMapping("/registration/add-id-card")
-    public String documentStageRegistrationAddIdCard(Model model) {
-        model.addAttribute("idcard", new IdentificationCard());
-        return "registration/document_stage";
-    }
-    @GetMapping("/registration/add-dl")
-    public String documentStageRegistrationAddDL(Model model) {
-        model.addAttribute("dl", new DriversLicense());
-        return "registration/document_stage";
-    }
-    @GetMapping("/registration/add-passport")
-    public String documentStageRegistrationAddPassport(Model model) {
-        model.addAttribute("idcard", new Passport());
-        return "registration/document_stage";
-    }
-
-    @PostMapping("/registration/add-id-card")
-    public String documentStageRegistrationAddIdCard(@Valid IdentificationCard card, BindingResult result, Principal principal) {
-        return "redirect:/";
-
-    }
-    @PostMapping("/registration/add-dl")
-    public String documentStageRegistrationAddDL(@Valid DriversLicense dl, BindingResult result, Principal principal) {
-        return "redirect:/";
-    }
-    @PostMapping("/registration/add-passport")
-    public String documentStageRegistrationAddPassport(@Valid Passport passport, BindingResult result, Principal principal) {
-        return "redirect:/";
-    }
